Remove local broadcast receiver to prevent run exceptions. Added option to set tile for extension update.
