+++ app/src/main/java/eu/kanade/tachiyomi/data/notification/NotificationHandler.kt
@@ -3,6 +3,7 @@ package eu.kanade.tachiyomi.data.notification
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
+import android.net.Uri
 import eu.kanade.tachiyomi.ui.main.MainActivity
 import eu.kanade.tachiyomi.util.getUriCompat
 import java.io.File
@@ -43,11 +44,10 @@ object NotificationHandler {
      * Returns [PendingIntent] that prompts user with apk install intent
      *
      * @param context context
-     * @param file file of apk that is installed
+     * @param uri uri of apk that is installed
      */
-    fun installApkPendingActivity(context: Context, file: File): PendingIntent {
+    fun installApkPendingActivity(context: Context, uri: Uri): PendingIntent {
         val intent = Intent(Intent.ACTION_VIEW).apply {
-            val uri = file.getUriCompat(context)
             setDataAndType(uri, "application/vnd.android.package-archive")
             flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_GRANT_READ_URI_PERMISSION
         }
