+++ app/src/main/java/eu/kanade/tachiyomi/data/updater/GithubRelease.kt
@@ -11,8 +11,8 @@ import com.google.gson.annotations.SerializedName
  * @param assets assets of latest release.
  */
 class GithubRelease(@SerializedName("tag_name") val version: String,
-        @SerializedName("body") val changeLog: String,
-        @SerializedName("assets") val assets: List<Assets>) {
+                    @SerializedName("body") val changeLog: String,
+                    @SerializedName("assets") private val assets: List<Assets>) {
 
     /**
      * Get download link of latest release from the assets.
