+++ app/src/main/java/eu/kanade/tachiyomi/data/updater/GithubUpdateResult.kt
@@ -3,5 +3,5 @@ package eu.kanade.tachiyomi.data.updater
 sealed class GithubUpdateResult {
 
     class NewUpdate(val release: GithubRelease): GithubUpdateResult()
-    class NoNewUpdate(): GithubUpdateResult()
+    class NoNewUpdate : GithubUpdateResult()
 }
