+++ app/src/main/java/eu/kanade/tachiyomi/util/FileExtensions.kt
@@ -13,9 +13,8 @@ import java.io.File
  * @param context context of application
  */
 fun File.getUriCompat(context: Context): Uri {
-    val uri = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N)
+    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N)
         FileProvider.getUriForFile(context, BuildConfig.APPLICATION_ID + ".provider", this)
     else Uri.fromFile(this)
-    return uri
 }
