+++ app/src/main/res/values/strings.xml
@@ -447,4 +447,5 @@
     <string name="channel_common">Common</string>
     <string name="channel_library">Library</string>
     <string name="channel_downloader">Downloader</string>
+    <string name="channel_updater">Updater</string>
 </resources>
